==> Revisao_Literatura/medium/3 - Parser Articles.ipynb <==
import pandas as pd
from tqdm.notebook import tqdm
import numpy as np 
from pathlib import Path
from bs4 import BeautifulSoup
tqdm.pandas()

path = 'data'

root_folder = Path(path)
root_folder.mkdir(parents=True, exist_ok=True)

df_searchterms = pd.read_csv(root_folder /'medium_search_base_detail.csv', sep = ';')

df_searchterms['path_txt_folder'] = path + '\\articles_txt\\'+ df_searchterms['folder']
df_searchterms['path_txt_article'] = df_searchterms['path_txt_folder']+ '\\'+df_searchterms['algoliaObjectId']+'.txt'

df_searchterms.sort_values(by = ['folder','voterCount','mediumUrl'],\
                            ascending=[False,False,False], ignore_index=True, inplace=True)
df_searchterms.drop_duplicates(subset='algoliaObjectId',keep='first', inplace = True)

df_searchterms.to_csv(root_folder /'medium_search_base_detail_txt.csv', sep = ';', index = False)

records = df_searchterms.to_dict(orient = 'records')

for record in tqdm(records):


    full_path_folder = Path(record['path_txt_folder'])
    full_path_folder.mkdir(parents=True, exist_ok=True)

    full_path_article = Path(record['path_file'])
    full_path_txt = Path(record['path_txt_article'])

    if full_path_txt.exists() and full_path_txt.is_file():  
        continue

    if not(full_path_article.exists() and full_path_article.is_file()):
         continue

    
    f = open(str(full_path_article),"r", encoding="utf-8")
    soup = BeautifulSoup(f.read(), "html.parser")
    f.close()
    section = soup.find('section')
    if not(section):
        continue

    content = section.get_text(separator='\n')
    
    with open(full_path_txt, "w", encoding="utf-8") as f:
        f.write(content)
